+++ Mountain-UI-Companion/AppDelegate.swift
@@ -6,6 +6,7 @@
 //
 
 import UIKit
+
 import Amplify
 import AWSAPIPlugin
 import AWSCognitoAuthPlugin
@@ -13,83 +14,32 @@ import AWSS3StoragePlugin
 
 @main
 class AppDelegate: UIResponder, UIApplicationDelegate {
-
-
-
+    
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey: Any]?) -> Bool {
         do {
-            try Amplify.add(plugin: AWSAPIPlugin(modelRegistration: AmplifyModels()))
             try Amplify.add(plugin: AWSCognitoAuthPlugin())
+            try Amplify.add(plugin: AWSAPIPlugin(modelRegistration: AmplifyModels()))
             try Amplify.add(plugin: AWSS3StoragePlugin())
             try Amplify.configure()
+            Swift.debugPrint("Amplify configured with auth plugin")
         } catch {
-            print("An error occurred setting up Amplify: \(error)")
+            Swift.debugPrint("Failed to initialize Amplify with \(error)")
         }
         return true
     }
-
+    
     // MARK: UISceneSession Lifecycle
-
+    
     func application(_ application: UIApplication, configurationForConnecting connectingSceneSession: UISceneSession, options: UIScene.ConnectionOptions) -> UISceneConfiguration {
         // Called when a new scene session is being created.
         // Use this method to select a configuration to create the new scene with.
         return UISceneConfiguration(name: "Default Configuration", sessionRole: connectingSceneSession.role)
     }
-
+    
     func application(_ application: UIApplication, didDiscardSceneSessions sceneSessions: Set<UISceneSession>) {
         // Called when the user discards a scene session.
         // If any sessions were discarded while the application was not running, this will be called shortly after application:didFinishLaunchingWithOptions.
         // Use this method to release any resources that were specific to the discarded scenes, as they will not return.
     }
     
-    func signUp(username: String, password: String, email: String) async {
-        let userAttributes = [AuthUserAttribute(.email, value: email)]
-        let options = AuthSignUpRequest.Options(userAttributes: userAttributes)
-        do {
-            let signUpResult = try await Amplify.Auth.signUp(
-                username: username,
-                password: password,
-                options: options
-            )
-            if case let .confirmUser(deliveryDetails, _, userId) = signUpResult.nextStep {
-                print("Delivery details \(String(describing: deliveryDetails)) for userId: \(String(describing: userId))")
-            } else {
-                print("SignUp Complete")
-            }
-        } catch let error as AuthError {
-            print("An error occurred while registering a user \(error)")
-        } catch {
-            print("Unexpected error: \(error)")
-        }
-    }
-    
-    func confirmSignUp(for username: String, with confirmationCode: String) async {
-        do {
-            let confirmSignUpResult = try await Amplify.Auth.confirmSignUp(
-                for: username,
-                confirmationCode: confirmationCode
-            )
-            print("Confirm sign up result completed: \(confirmSignUpResult.isSignUpComplete)")
-        } catch let error as AuthError {
-            print("An error occurred while confirming sign up \(error)")
-        } catch {
-            print("Unexpected error: \(error)")
-        }
-    }
-    
-    func signIn(username: String, password: String) async {
-        do {
-            let signInResult = try await Amplify.Auth.signIn(
-                username: username,
-                password: password
-                )
-            if signInResult.isSignedIn {
-                print("Sign in succeeded")
-            }
-        } catch let error as AuthError {
-            print("Sign in failed \(error)")
-        } catch {
-            print("Unexpected error: \(error)")
-        }
-    }
 }
